refactor: EventType 관련 코드를 type 디렉토리로 이동
