+++ module-api/src/main/java/com/trybe/moduleapi/challenge/event/listener/PopularChallengeEventListener.java
@@ -1,7 +1,7 @@
 package com.trybe.moduleapi.challenge.event.listener;
 
 import com.trybe.moduleapi.challenge.event.model.ChallengeEvent;
-import com.trybe.moduleapi.challenge.event.ChallengeEventType;
+import com.trybe.moduleapi.challenge.event.type.ChallengeEventType;
 import com.trybe.moduleapi.utils.DateUtils;
 import com.trybe.modulecore.challenge.repository.popular.PopularChallengeCache;
 import org.springframework.scheduling.annotation.Async;
